requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
backports.zoneinfo==0.2.1
Django==4.1.6
djangorestframework==3.14.0
pytz==2022.7.1
redis==4.5.0
sqlparse==0.4.3
django
markdown
django-filter
psycopg2-binary